requirements.txt
tensorflow>=2.12.0
streamlit>=1.25.0
pandas>=1.5.0
numpy>=1.23.0
opencv-python-headless>=4.7.0.72
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.2
scikit-learn>=1.2.0
plotly>=5.14.0
fpdf2>=2.7.4
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
icrawler>=0.6.7
